baselines/gloria/requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
albumentations==0.5.2
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.0
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
decorator==4.4.2
docker-pycreds==0.4.0
efficientnet-pytorch==0.6.3
filelock==3.0.12
fsspec==0.8.5
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.1
idna==2.10
imageio==2.9.0
imgaug==0.4.0
joblib==1.0.0
kaggle==1.5.10
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
multidict==5.1.0
munch==2.5.0
networkx==2.5
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.5
opencv-python==4.5.1.48
opencv-python-headless==4.5.3.56
packaging==20.8
pandas==1.2.0
pathtools==0.1.2
Pillow==8.1.0
pretrainedmodels==0.7.4
promise==2.3
protobuf==3.14.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydicom==2.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-slugify==4.0.1
pytorch-lightning==1.1.4
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
sacremoses==0.0.43
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
segmentation-models-pytorch==0.2.0
sentry-sdk==0.19.5
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.1.14
timm==0.4.12
tokenizers==0.9.4
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
transformers==4.2.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wandb==0.10.14
watchdog==0.10.4
Werkzeug==1.0.1
yarl==1.6.3
